Employee Management System
The purpose of “Employee Management System” is to automate the existing system by the help of computerized equipment’s and full-fledged computer software, fulfilling their requirements so that valuable information get stored for longer time period. It can lead to error free, secure, reliable, fast management system which assist the user in easy record keeping, so it proves it is user friendly. It manages the details of employee. Editing, adding, updating of records is improved which results in proper resource management of employee data